requirements.txt
# Appium-Python-Client==0.52;python_version < '3.0'
# Appium-Python-Client==1.0.2;python_version >= '3.0'

Appium-Python-Client==0.52;python_version < '3.0'
Appium-Python-Client>=2.7.1;python_version >= '3.0'
selenium>=4.0.0